[add] slot execution logging
The saving backend machinery was extended:
- now there's a SavingBackend class instead of a method, because we
need two methods, one to save emitted signals, another to save
executed slots;
- there is an interface and three implementations: an empty one which
does nothing, another one which dumps things to the system log, and
the third one which saves things to the app database;
- for the latter one to work, we create a new table for
SlotExecutionLog entity (a table for EmittedSignals is the same as
before).

For a slot execution, we record whether it succeeded or failed (and
the error reason if it did), slot arguments, an id of the signal that
triggered it, and the time passed from launching the slot to its
finish (it's not a pure processor time since the slot may contain
awaits itself, but that's what we have).

Exhaustive tests for backends and for the bus with backends were added.